requirements.txt
# Python packages needed for helper scripts
python-dotenv>=1.0
